# feat: implement cached item services with monitoring