--- a/.github/workflows/deploy.yml
+++ b/.github/workflows/deploy.yml
@@ -42,25 +42,21 @@ jobs:
           tags: |
             ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}:latest
             ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}:${{ github.sha }}
+          labels: |
+            org.opencontainers.image.source=${{ github.repositoryUrl }}
 
       - name: Deploy to Azure Web App
         uses: azure/webapps-deploy@v2
         with:
           app-name: ${{ secrets.AZURE_WEBAPP_NAME }}
           publish-profile: ${{ secrets.AZURE_PUBLISH_PROFILE }}
-          images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}@${{ steps.build-push.outputs.digest }}
+          images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}:${{ github.sha }}
+          runtime: "linux"
 
       - name: Log in to Azure (for app service settings)
         uses: azure/login@v1
         with:
           creds: ${{ secrets.AZURE_CREDENTIALS }}
-        # AZURE_CREDENTIALS should be a JSON string for a Service Principal:
-        # {
-        #   "clientId":   "<GUID>",
-        #   "clientSecret": "<SECRET>",
-        #   "subscriptionId": "<GUID>",
-        #   "tenantId": "<GUID>"
-        # }
 
       - name: Configure environment variables
         uses: azure/appservice-settings@v1