--- a/API-Core/.env.example
+++ b/API-Core/.env.example
@@ -28,6 +28,12 @@ PORT=5432
 DBNAME=postgres
 
 # Full Supabase URI (use URL-encoded password here)
-# You can encode password using:
-# >>> import urllib.parse; urllib.parse.quote("your-password")
+#>>>>>>>>>
+    # You can encode password using:
+    # >>> import urllib.parse; urllib.parse.quote("your-password")
+#<<<<<<<<<
 SUPABASE_CONN="postgresql://<USER>:<URL_ENCODED_PASSWORD>@${HOST}:${PORT}/${DBNAME}"
+
+#caching
+REDIS_URL=redis://localhost:6379/0
+CACHE_ENABLED=true