… API-Core/app/middleware/security.py
… API-Core/app/services/enhanced_item.py